+++ src/Trie.h
@@ -10,7 +10,7 @@
 struct Uzel {
   bool is_terminal;
   std::vector<int32_t> next;
-  std::vector<int32_t> code;
+  std::vector<int8_t> code;
   Uzel(bool is_end) {
     is_terminal = is_end;
     this->next.resize(1 << CHAR_SIZE);
@@ -28,7 +28,7 @@ public:
   Trie();
   ~Trie();
   void insert(const std::string& s);
-  void insert(const std::string& s, const std::vector<int32_t>& code);
+  void insert(const std::string& s, const std::vector<int8_t>& code);
   bool is_next(int32_t uz, unsigned char c) const;
   int32_t next(int32_t uz, unsigned char c) const;
 };
